img-caption project/requirements.txt
transformers==4.28.0
torch==2.1.0
Pillow==9.3.0
tkinter==8.6 (موجود بشكل افتراضي في بايثون)
nltk==3.7
